[lib] Accept keyserver ID in blob service upload action
Summary:

Due to changes in D9399 the blob service upload action now requires `threadID` to extract keyserver ID from it. It is needed to determine which keyserver we want to send upload metatata to.

The thread ID is unavailable in some context, e.g. for user avatars. However, the extraction function is a no-op if we passed keyserver ID already so renamed the input argument to indicate that.

More context in [[ https://linear.app/comm/issue/ENG-5673/useblobserviceupload-requires-threadid-which-is-not-available-for-user | ENG-5673 ]].

Test Plan:

- Created another unit test for `extractKeyserverIdFromID()`
- Flow passes
- Multimedia uplaods still work
- Avatars tested together with other diffs in this stack (two child diffs)

Reviewers: inka!, michal, tomek, ashoat

Subscribers: