[lib] Move validators to specs
Summary:
Move the validators. Use the same type for client and server updates - they are the same. In the future we can consider merging the types, but that would require updating types in a lot of places. Also, not sure if we need that, because in the future the update types might become different again - in such a case we would need to introduce a new field in a spec to differentiate between client and server validators.

https://linear.app/comm/issue/ENG-4239/add-validators-to-update-specs

Test Plan: Flow, validation tests

Reviewers: kamil, inka, bartek

Reviewed By: kamil

Subscribers: ashoat, wyilio

Differential Revision: https://phab.comm.dev/D9741