[CI] Install python 3.10 for macos
Summary:
[ENG-5678](https://linear.app/comm/issue/ENG-5678)

The general error that breaks the github CI is the same as in this [issue](https://github.com/electron/forge/issues/2807) - `distutils` was removed in a newer version of python and so one of the dependencies postinstall script breaks because of that. This diff fixes that by additionaly installing an older (3.10) version of python before installing the dependencies.

Note that I'm still not sure why the CI is sometimes successful but I'm hoping this should fix the errors.

INFO: Most native (written in C++) js modules are built with something called `node-gyp`. It's written in python and used at build time in postinstall scripts. That's why we need to mess with python.

Test Plan: Tested on a GitHub fork. Without this changes CI failed. After this changes tried running it 3 times and it succeded every time.

Reviewers: atul, ashoat, tomek

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9760