[lib/web/native] Show voiced_in_announcement_channels user-surfaced permission for all community types
Summary:
When filtering out the user-surfaced 'voiced in announcement channels' permission when creating/editing roles, I didn't consider that we should always display this option because subchannels can be announcement threads. Now that they can be, I'm not going to filter this option out anymore.

Once [[ https://linear.app/comm/issue/ENG-5183/voiced-in-announcement-channels-doesnt-seem-to-have-the-right | ENG-5183 ]] is resolved, this will mean that ticking this checkbox will allow users with the role to able to talk in all announcement channels (including the community root if it's an announcement thread).

This shouldn't require a migration since all it does is display the option in the permissions list client-side to be able to be toggled.

Resolves https://linear.app/comm/issue/ENG-5182/cant-see-voiced-in-announcement-channels-permission-when-editing-role

Depends on D9608

Test Plan:
Please see the testing video below where I confirm in both announcement and non-announcement communities, the user-surfaced permission 'voiced in announcement channels' is shown:

{F838658}

Reviewers: atul, ginsu, ashoat

Reviewed By: atul

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9514