[lib/web/native] Show voiced_in_announcement_channels user-surfaced permission for all community types
Summary:
When filtering out the user-surfaced 'voiced in announcement channels' permission when creating/editing roles, I didn't consider that we should always display this option because subchannels can be announcement threads. Now that they can be, I'm not going to filter this option out anymore. This will mean that ticking this checkbox, users with this role will be able to talk in all announcement channels (including the community root if it's an announcement thread).

This shouldn't require a migration since all it does is display the option in the permissions list client-side to be able to be toggled.

Resolves https://linear.app/comm/issue/ENG-5182/cant-see-voiced-in-announcement-channels-permission-when-editing-role

Depends on D9489

Test Plan:
Please see the test video below where I do the following:

1. Created a new community and add some members (making it an announcement root)
2. Create some normal subchannels and announcement subchannels
3. Log out and confirm that an added user cannot speak in the announcement subchannels but can speak in the normal subchannels
4. Log back into the Admin user and modify the `Members` permissions so they can speak in announcement channels
5. Log into the normal members and confirm that they can talk in the announcement root + announcement subchannels
6. Log into the Admin and create a custom role without 'voiced_in_announcement_channels' permission and assign it to some members
7. Log into the members and confirm that they can no longer speak in the announcement root and announcement subchannels
8. Log back into the Admin user and edit the new role to have 'voiced_in_announcement_channels' permission
9. Log into the other members and confirm they can now speak in announcement subchannels and the announcement root

Reviewers: atul, ginsu, ashoat

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9514