[native] introduce keyserver selection bottom sheet ui
Summary:
This diff introduces the keyserver selection bottom sheet ui.

Please note that the focus of this diff was to introduce the UI. Subsequent diffs will polish/finish up the keyserver selection bottom sheet remove functionality, fix the snap points so that there is not all this blank space underneath the content of the bottom sheet, handling keyservers that are not connected, etc.

Linear task: https://linear.app/comm/issue/ENG-5469/introduce-keyserverselectionbottomsheet-component-ui

For context here is a screenshot of the designs:
{F864064}

Depends on D9753

Test Plan:
Please see the screenshots below

{F864196}

Reviewers: inka, rohan, michal

Reviewed By: inka

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9754