[lib] Introduce `decodeMinimallyEncodedRawThreadInfo`
Summary:
This allows us to "recover" `RawThreadInfo` from `MinimallyEncodedRawThreadInfo` encoded via `minimallyEncodeRawThreadInfo`.

This is *finally* the last utility function that we'll be introducing in this stack. We now have all of the building blocks in place and can make necessary changes to `keyserver`/`native`/`web`.

This function is also "temporary" because in the future `MinimallyEncodedRawThreadInfo` will effectively be the new `RawThreadInfo`. Usage of this function will be replaced by refactoring client code to query the minimally encoded permissions directly.

---

Depends on D9752

Test Plan: Unit test, will add more unit tests at end of stack with "prod data" from my Redux store.

Reviewers: ashoat, ginsu, tomek, rohan

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9755