[Flow202][web] [4/x] Extract definition of FetchRevisionsForEntryPayload
Summary: This is actually a `lib` change, but it helps since we need this type for a typehint to address a Flow error in the new version.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9769

Test Plan: Confirm no changes in Flow error count

Reviewers: tomek

Subscribers: