[lib] Inline reduceLastCommunicatedPlatformDetails
Summary:
issue: https://linear.app/comm/issue/ENG-5353/inline-reducelastcommunicatedplatformdetails
`updateLastCommunicatedPlatformDetailsActionType` is dispatched from socket and callServerEndpoint. In both of these we are in the context of one keyserver, so this action is keyserver specific

Test Plan:
called `updateLastCommunicatedPlatformDetailsActionType`, checked that it sets the lastCommunicatedPlatformDetails correctly, chcecked that `logInActionTypes` sets it correctly.

Reviewers:
michal, kamil, ginsu
Subscribers: