[keyserver] Update changeRoleThreadQuery to use the Voiced role for announcement channels
Summary:
Following introducing a new role for announcement subchannels, the first step I'd like to take is to give everyone added to the channel the `Voiced` role. This is not permanent, and will not be landed as-is, but it's an intermediate step in separating out `Members` and `Voiced`.

In `changeRole`, `changeRoleThreadQuery(threadID, role)` is called to get the intended role and permissions for the users being added to the thread. Naturally, this is the `default_role` column in MariaDB. We'll want to do some updates here to ultimately decide whether Voiced or Members is the appropriate role.

In this task, I'll make it so every new member gets `Voiced` instead of `Members`, and a later task will update this to check whether the user has 'Voiced in announcement channels' permission, and assign them a role accordingly.

Addresses [[ https://linear.app/comm/issue/ENG-5635/update-changerolethreadquery-to-use-the-voiced-role-for-announcement | ENG-5635 ]]

Depends on D9825

Test Plan:
Tested this in three ways:
- Confirmed that new users added to the announcement subchannel had `voiced` permissions in the `memberships` table in MariaDB
- Confirmed that new users could automatically speak in announcement subchannels (since they have the `Voiced` role)
- Changed the member pill that shows the role to use the current `threadInfo` (as opposed to the community thread info) just to see that:
-- in announcement subchannels, new members added had the `Voiced` role
-- in normal subchannels, new members added had the `Members` role still

Reviewers: atul, ginsu, ashoat

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9805