[native] calculate snap points for keyserver selection bottomsheet
Summary: In order to correctly get our bottomsheet to snap up to the correct height we need to measure the container of the content. When I was working on user profiles, I introduced a bottomsheet context that exposed a `setState` for the bottomsheet height. We can reuse that here.

Test Plan: Please see the demo videos below

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9795