[Flow202][keyserver] [1/x] Fix Flow issues with conditional changeRole passed to promiseAll
Summary:
The new Flow is much stricter about `promiseAll`. It's relaxed a bit in Flow 0.214, but for now we have to comply with the additional strictness.

I separated out this diff because it required some minimal changes to `thread-permission-updates.js` in order to rename and export a type. This way the context for those changes is more clear.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9786

Test Plan: Confirm the Flow errors go away

Reviewers: tomek