[lib] Just cut/paste values into `minimallyEncoded` validators
Summary:
Changes in D9808 did not resolve Jest issue.

Spent way too much time trying to get this to work, and the core issue is that we're spreading in base validators...which is such a minor thing so I'm just going to cut/paste them for now to unblock.

---

Depends on D9808

Test Plan: Tests continue to pass, Jest issues are no longer a problem.

Reviewers: ashoat, ginsu, tomek, rohan

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9809