[native] VersionSupportedChecker component
Summary:
this component calls the identity service and displays an alert if the client code version is not supported by the identity service.

Depends on D9813

Test Plan: built the native app with an unsupported code version and saw the alert. built with the actual code version and no alert was displayed.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: inka, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9814