[lib/native] extend keyserverAdminUsername to be keyserverAdminUserInfo
Summary: The remove keyserver action needs the keyserver admin user id. I realized as I was implementing the solution to remove a keyserver from the keyserver store that I am only had the username available to me. This diff extends the `keyserverAdminUsername` to be `keyserverAdminUserInfo` which also holds the keyserver admin user id

Test Plan: `flow` and confirmed that I now have the userID in the `keyserverAdminUserInfo`

Reviewers:

Subscribers: