[lib] Inline reduceConnectionInfo pt.2
Summary:
issue: https://linear.app/comm/issue/ENG-5352/inline-reduceconnectioninfo
See the deleted file for how it worked before
The disconnected bar and with it the `lateResponses` logic will be reviewd again during ENG-4126 and ENG-4125. For now I'm refactoring them in a way that 1. doesn't break the old behaviour 2. Will be easier to refactor later on

Test Plan: Dispatched setNewSessionActionType, updateDisconnectedBarActionType, logOutActionTypes, setLateResponseActionType actions and checked that the state changes as expected

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9815