[keyserver] Change updateRoles to support Voiced role
Summary:
`updateRoles` needs to be changed to support the `Voiced` role when a thread's type is changed.

For converting a channel from a regular channel to an announcement channel, we need to first create a `Vocied` role. Then, we need to query the roles in the community channel that have `voiced_in_announcement_channels`, and if the member has that role give them `Voiced`, otherwise give them `Members`.

For converting a channel from an announcement channel to a regular channel, I can take advantage of the `deleteRole` function I created that will both delete the role and assign the affected users to the default role of the thread (without triggering any messages in chat).

Addresses [[ https://linear.app/comm/issue/ENG-5702/change-updateroles-to-support-voiced | ENG-5702 ]]

Depends on D9827

Test Plan:
Converting a regular channel to an announcement channel:
- Created a regular subchannel with four members
- Created a community role that had `VOICED_IN_ANNOUNCEMENT_CHANNELS`
- Assigned it to two members
- Ran `yarn script dist/scripts/make-channel-private.js` to convert the channel to `type: threadTypes.COMMUNITY_ANNOUNCEMENT_OPEN_SUBTHREAD`
- Confirmed that the `if (rolePermissions.Voiced  &&  !currentRolePermissions.Voiced) {` condition was called
- Confirmed that a `Voiced` role was created
- Confirmed that the users who had the community role with `VOICED_IN_ANNOUNCEMENT_CHANNELS` had their channel role set to the `Voiced` role. The other users remained with the `Members` role

Converting an announcement channel to a regular channel:
- Patched the changes from D9805 that makes new members added to an announcement channel have the 'Voiced' role
- Created an announcement channel and added four members
- Confirmed that each of these members had the `Voiced` role in the `memberships` table
- Ran `yarn script dist/scripts/make-channel-private.js` to convert the channel to `type: threadTypes.COMMUNITY_OPEN_SUBTHREAD`
- Confirmed that the `else  if (!rolePermissions.Voiced  && currentRolePermissions.Voiced) {` condition was called
- Confirmed that the `Voiced` role was deleted
- Confirmed that the `memberships` table was updated so the four members who previously had `Voiced` were assigned the default role for the channel (`Members`)

Reviewers: atul, ginsu, ashoat

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9833