[identity] Add RPC for finding user ID
Summary:

[[ https://linear.app/comm/issue/ENG-4132/identity-service-rpc-to-check-if-a-username-exists | ENG-4132 ]]. Added RPC to find user ID by either username or wallet address.
The API is the same as for GetOutboundKeys RPC, which will then be refactored to instead accept user ID.

Test Plan:

Created a test user, used BloomRPC to call the RPC with the username. It returned the user ID - confirmed with DDB. A unit test is added in the next diff.

Reviewers: varun, ashoat, michal

Subscribers: