[Flow202][keyserver] [11/x] Address type errors with memberOfContainingThread
Summary:
The new version of Flow caught a type error here. `memberOfContainingThread` should be a `boolean`, but can be set to be nullable. This diff casts it to a `boolean` to avoid the type error.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9859

Test Plan: Confirm the Flow errors go away

Reviewers: tomek

Subscribers: