[Flow202][keyserver] [12/x] Address type errors with getInitialReduxStateResponder
Summary:
The new version of Flow seems to be worried that the value of `finalNavInfo.activeChatThreadID` will change between when it is being checked and when it's used. This diff assigns the value to a local variable before checking it, which Flow seems to prefer.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9860

Test Plan: Confirm the Flow errors go away

Reviewers: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9861