[web] remove `CommQueryExecutor` code
Summary:
It was introduced to test WASM on the web as entry point some time ago, right now we can remove it.

One additional thing we could consider is that we still refer to `.wasm` file as `commQueryExecutorFilename` while now it's `SQLiteQueryExecutor`, but not sure if changing it worth now.

Test Plan:
1. Flow
2. Build Emscripten
3. Make sure there is no errors on worker console

Reviewers: tomek, michal

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10204