[lib] Introduce `MinimallyEncodedRelativeMemberInfo` type
Summary:
We're introducing `MinimallyEncodedRelativeMemberInfo` which will be a field of the soon-to-be-introduced `MinimallyEncodedThreadInfo`.

We've refactored most of the utils in `thread-utils` that accepted `RawThreadInfo` to also accept `MinimallyEncodedRawThreadInfo`. We will now introduce `MinimallyEncodedThreadInfo` and update `thread-utils` and beyond to support it where `ThreadInfo` is supported today.

---

Depends on D9842

Test Plan: I'm NOT going to introduce unit tests for all of these type utilities like I did earlier in the stack. I'm going to stick to trusting Flow for now since I think benefit of adding unit tests is questionable. Will generally be adding unit tests at end of stack, so that might be opportunity to write tests for these.

Reviewers: ashoat, ginsu, tomek, rohan

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D9843