[lib] Introduce `MinimallyEncodedThreadInfo` type
Summary:
We need corresponding `MinimallyEncodedThreadInfo` for `MinimallyEncodedRawThreadInfo`. The only really relevant difference between the two is that `MinimallyEncodedThreadInfo` has `$ReadOnlyArray<MinimallyEncodedRelativeMemberInfo>` in `members` field instead of `$ReadOnlyArray<MinimallyEncodedMemberInfo>`.

---

Depends on D9845

Test Plan: Flow. Earlier in the stack I was writing unit tests for all of these, but I think the value of unit tests for these is questionable. Will defer for now and consider writing at the end.

Reviewers: ashoat, ginsu, tomek, rohan

Differential Revision: https://phab.comm.dev/D9846