[keyserver][lib] Update freshMessageStore to return currentAsOf for all keyservers
Summary:
issue: https://linear.app/comm/issue/ENG-5722/update-freshmessagestore-function
`freshMessageStore` should return `currentAsOf` for all keyservers

Test Plan:
Tested that after the `setInitialReduxState` action the `currentAsOf` field is correct - a "map" from keyserver id to number.
This tests this function, because `freshMessageStore` is called by the keyserver in `getInitialReduxStateResponder` which returns the initial state. The states from all keyservers are then merged in `getInitialReduxState`, and the returnd value is dispatched with `setInitialReduxState` action.
I also logged to the console the value returned from `freshMessageStore` in `getInitialReduxStateResponder` and checked that it is correct.
I also logged in and checked that the `currentAsOf` in `messageStore` is correct after `logInActionTypes.success`

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9879