[web/native] Use useSearchUsers for the friends list
Summary:
`useSearchUsers` is used to include user search results from the server (outside of the user's local userStore during the 'composing a new chat' flow on both web and native.

The code looks a little like the code in `AddUsersList` on web and `RelationshipList` on native when viewing your friends. Because of this, this diff is really useful in getting rid of a 'one off' search logic
and instead using a shared search helper. This will make it easier to do the next part of this stack (updating search helpers to lookup addresses from a typed out ENS name).

Addresses [[ https://linear.app/comm/issue/ENG-5490/try-to-use-usesearchusers-for-the-friends-list | ENG-5490 ]]

Depends on D9883

Test Plan: Confirmed that searching for users outside of the user's local userStore still worked in these two search experiences. Made sure to compare the search experience to how it is on production right now as well to make sure no regressions were introduced.

Reviewers: atul, ginsu, inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9884