[Flow202][native][skip-ci] [1/x] Ignore electron-update-server in native
Summary:
Not sure why the new Flow is seeing this directory while the old did not. However, we've had this problem before, as we can see with `native/.flowconfig` ignoring a variety of other projects.

(The reason we need this is that the Flow in `native` doesn't have libdefs set up for eg. Electron.)

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9751

Test Plan: Confirm the Flow errors go away

Reviewers: michal

Reviewed By: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9985